requirements.txt
deepchem>=2.8.0
rdkit>=2026.3.1
torch>=2.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0.0
